change appcenter sh to rb file
